feat: add update product

+++ src/application/usecases/product/delete-product.ts
@@ -1,9 +1,12 @@
 import type { ProductRepository } from '@/application/ports/product-repository'
+import { NotFoundError } from '@/domain/errors/not-found-error'
 
 export class DeleteProduct {
 	constructor(private readonly productRepository: ProductRepository) {}
 
 	async execute(id: string): Promise<void> {
+		const product = await this.productRepository.findById(id)
+		if (!product) throw new NotFoundError('Product not found')
 		await this.productRepository.delete(id)
 	}
 }
